requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-admin-list-filter-dropdown==1.0.2
django-ckeditor==5.7.1
django-cleanup==4.0.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.0
django-js-asset==1.2.2
django-modeltranslation==0.13.3
django-mptt==0.9.0
django-querysetsequence==0.11
django-recaptcha==2.0.5
django-recaptcha2==1.4.1
django-simple-metatags==0.9.2
django-suit==2.0a1
html5lib==1.0.1
idna==2.8
ipaddress==1.0.22
more-itertools==7.2.0
Pillow==6.1.0
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2019.2
reportlab==3.5.23
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
transliterate==1.10.2
urllib3==1.25.3
webencodings==0.5.1
xhtml2pdf==0.2.3
